pyproject.toml
[tool.poetry]
name = "meeting-notes-ai"
version = "0.1.0"
description = ""
authors = ["jkarenko <juho.karenko@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.8.0"
fastapi = "^0.109.0"
ffmpeg-python = "^0.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
